Trac #29672: detect the palp class{,-4d}.x executable.
There are two possible names for the "class-4d.x" executable used in
the palp_database.py file. The upstream name for it is simply
"class.x", but SageMath has historically installed an optimized
version as "class-4d.x" and several distributions have followed
suit. This commit updates the spkg-configure.m4 for palp to detect
which the best available name, and substitute it into the environment
variable SAGE_PALP_CLASS_EXE.

Note that this executable is only used in optional functionality
that requires an 8.6GiB old-style SPKG database. In other words,
it is untested.