Trac #29729: add special case to solve_right() for symbolic systems.
Elements of SR "remember" whether or not they are exact, and if we are
given a system consisting entirely of exact elements, then the "check"
need not be elided for want of exactitude. This commit adds a special
case to the solve_right() method for matrices to handle SR with care.