Trac #33159: add special case to solve_right() for symbolic systems.
Elements of SR remember whether or not they are exact, and if we are
given a system consisting entirely of exact elements, then the checks
for correctness in solve_left() and solve_right() need not be skipped
as they are for other inexact rings. This commit adds a special case
to the solve_right() method for matrices to handle SR with care, and
includes an additional test case from Trac 29729. Since solve_left()
is implemented in terms of solve_right(), both are addressed.