Trac #29493: add sanity checks for GLPK's variable bound methods.
A crash was reported in Trac #10232, affecting in particular the
variable_lower_bound() and variable_upper_bound() methods of the GLPK
backend. These crashes were fixed by wrapping them in sig_on() and
sig_off(), and then throwing a GLPKError in the event that GLPK
crashes.

That approach is slightly perverse: in both cases, we know what
variable indices are valid at the beginning of the function. This
commit adds a sanity check to those methods that throws a ValueError
if the variable index is out of bounds. This avoids crashing GLPK and
thus partially eliminates our reliance on a custom GLPK error-handling
patch that is persona non grata.