Trac #29493: add bounds checks to GLPK's add_linear_constraint() method.
This commit ensures that all of the variable indices passed to GLPK's
add_linear_constraint() method are valid. This was partially addressed
in Trac #19525, but that approach requires us to let GLPK to first
crash on the invalid input. The new bounds check catches the mistake
before GLPK sees it, and now returns a ValueError mentioning the
invalid index rather than a GLPKError.