Trac #29493: add bounds checks to GLPK's get_col_dual() method.
The get_col_dual() method of the GLPK backend would crash if it
received an invalid column index, because that index is passed
directly to GLPK. I have added "except? -1" to the method signature,
because otherwise it (being a C method that returns a C type) would be
unable to raise a Python exception. The question mark is needed to
avoid a conflict with the otherwise-valid dedicated return value -1.
Afterwards, the added bounds check that raises a ValueError is
completely standard.