Trac #29493: verify input to GLPK row/column tableaux methods.
The eval_tab_row() and eval_tab_col() methods of the GLPK backend
require a basis factorization to exist, and the indicated variable to
be basic or non-basic, respectively. In the past we have relied upon
GLPK crashing to indicate that one of these conditions was not
met. This commit uses the existing is_variable_basic() and
is_slack_variable_basic() methods to verify that the given index is
(non-)basic, and uses a new GLPK function glp_bf_exists() to check
that the basis factorization exists. This should prevent GLPK from
crashing, and reduces our dependence on custom GLPK patches.