Trac #29493: make simplex_then_intopt the default GLPK method.
Much of the GLPK error handling that requires custom patching is due
to the tendency of the "glp_intonly" method to crash on problems it
can't solve. The upstream recommendation is to run any suspicious
problems through the (continuous) simplex solver first to e.g. detect
infeasibility. That process is already implemented in SageMath via
the "simplex_then_intopt" solver parameter.

This commit changes "simplex_then_intopt" to be the default. Users can
still set the parameter to "intonly" (or anything else) if they're
sure that the problem is solvable, but this commit also updates the
documentation with the relevant warnings against doing so blindly.