Adding Slackware-multilib to write-dockerfile
This allows portability testing on multilib Slackware systems (that is, the usual 64 bit with the added 32 bit capability). This system configuration is not official supported by Slackware, but is in common use. Because it required third party packages, I basically just added an extra if statement after inital package installation, but before bootstrapping.