Trac #30128: always source sage-env-config before sage-env.
The sage-env script needs to know the values of SAGE_ROOT and
SAGE_SCRIPTS_DIR, for example, but due to a chicken-and-egg problem it
can't source the sage-env-config script where they are defined. Since
whoever is sourcing sage-env perforce knows where it is located, this
commit updates those callers to first source sage-env-config. This
solves the inter-dependency, and does so in a POSIX-compatible way.

A future commit will remove the bashism in sage-env that tries to
determine the location of sage-env-config on its own.