Trac #30128: enforce sourcing of sage-env-config before src/bin/sage-env.
All callers within sage now source sage-env-config before sage-env, so
this commit removes the bashism within sage-env that tries to find and
source sage-env-config on its own. Since sage-env still morally
depends on sage-env-config for the values of some variables, we now
throw an error in sage-env is sage-env-config was not previously
sourced. This should prevent future misuse from creeping in.