29446: improve handling of SAGE_ROOT to avoid symbolic links
With this change, SAGE_ROOT will always be the physical path, even if
the current directory when compiling Sage is a symbolic link
(non-physical path) to Sage's root directory.

In particular, this fixes a test failure in src/sage/env.py when Sage
was built from the symbolic/non-physical path.