Trac #30153: factor out $(1)-no-deps targets in build/make/Makefile.in.
The no-deps package targets in build/make/Makefile.in are duplicates
of what I shall refer to as the "yes-deps" targets. Instead of
manually synchronizing the two which differ only in their
prerequisites, this commit changes the yes-deps targets to run $(MAKE)
on the no-deps targets after building the deps.