Trac #30200: add the global Niez1998 reference.
Add the paper "Group majorization and Schur type inequalities" by
Marek Niezgoda to src/doc/en/reference/references/index.rst, which
contains the global reference list. This paper mentions several
convex cones in the context of group majorization, and can be used
as a reference for some things in the cone catalogue.

Here's a BibTeX entry for good measure:

  @article{
    niez1998,
    author = "Marek Niezgoda",
    title = "Group majorization and {S}chur type inequalities",
    journal = "Linear Algebra and its Applications",
    volume = 268,
    number = 1,
    year = 1998,
    pages = "9--30",
    doi = "10.1016/S0024-3795(97)89322-6"
  }