Trac #30231: compute libpari_pari_cfg entirely in pre-post/check.
The post-check phase for gp2c's spkg-configure.m4 can fail if the
check for gp2c itself is never performed, because the value of the
variable libpari_pari_cfg never gets set to its default. Instead of
defaulting that variable in the check phase (based on the result of
the SAGE_SPKG_DEPCHECK macro), this commit does the needed stuffs
entirely in the pre- and post-check phases that are always performed,
even when the check itself is not.