Implement generic FinitelyGenerated axioms
This commit adds three axioms

1. `FinitelyGeneratedAsDistributiveMagma`. An object `C` with two magmatic
   operations `*` and `+`, such that `*` distributes over `+` is said to
   be finitely generated if there exists a finite subset `A \subset C`
   such that every element can be written as sum of elements of the form
   `a_1*...*a_n` with any parenthesis assignment and `a_i \in A` for
   every `i`
2. `FinitelyGeneratedAsAlgebra`. An object as `C` above with the
   which in addition is a `k`-module and the operation `*` is
   `k`-bilinear is finitely generated if every element can be written as
   a linear combination of elements as in 1.
3. `FinitelyGeneratedAsLieAlgebroid`. An object as `C` above, with an
   additional structure of being an `R`-module for a `k`-algebra `R`.
   `C` is called finitely generated if every element can be written as a
   `k`-linear combination of elements of the form
   `(r_1 a_1)*....*(r_n a_n)` with `a_i \in A` and `r_i \in R` for every
   `i` and any parenthesis assignment.

It adds a bare class `FinitelyGeneratedRings` and implements
`Rings().FinitelyGenerated()` and `LieAlgebras().FinitelyGenerated()`
just to show the difficulties that we would encounter with this generic
approach: for a Lie algebra, `gens` is not what is meant by "generators"
usually, but what is meant by a "basis". Of course finite `gens` implies
`FinitelyGeneratedAsAlgebra` but not conversely.

The same abuse is currently used in `LieConformalAlgebras` where `gens`
means a set of generators as an `R`-module (notation as in 3. above). So
again finite `gens` implies `FinitelyGeneratedAsAlgebroid` but not
conversely.