Trac #30533: add new -q (quiet) flag for the bootstrap process.
Most of the sage build system can be "quieted" in one way or another,
but our ./bootstrap script is an exception. To minimize the noise in
everyone's logs, this commit adds a "-q" flag (that stands for
"quiet") to the bootstrap script. For now, all it does is toggle
whether or not the bootstrap scripts themselves echo things to the
terminal. More work is needed to silence e.g. automake, which the
bootstrap script invokes.