ensure the order in the pythons list is respected
AC_PATH_PROGS_FEATURE_CHECK() has the outer loop over PATH
entries. Thus, a python on the list [python.foo python.baz ....] of
pythons that comes first in PATH and has all the features will be
picked, not the first python on the list.

We modify this macro to ensure sane order, by making the loop over
the list of pythons the outermost