Scale clipping planes by overall plot size to accomodate very large/small plots
Kept the same ratio between near and far clipping distances as before to
preserve current depth buffer precision, but skewed it bit more toward the near
clip plane for better zooming in. Can still zoom out plenty far. By the time
you hit the far plane, the plot is so tiny as to be useless anyways.