#18036 doctest updates: pari-gp interfaces
Note that gp(SR(I)).parent() is not SR but QQ[i]. This is consistent
with the fact that the gp-to-sage conversion currently just tries to
somehow make sense of the string representation of the Pari object.
Since QQ[i] coerces into SR, more complex symbolic expressions
containing I *will* be converted to elements of SR.

Also note that while SR(I) is converted to a Pari complex number,
I ∈ ℚ[i] itself is converted to a Pari number field element.