Trac #30921: fix verbose "epsilon" output in polynomial small_roots().
The verbose output from the small_roots() function reports its
"epsilon" parameter as an integer, but float values are possible. This
leads to incorrect output when, say, epsilon is "0.5" and "0" is
reported. This commit simply changes the guilty %d to an %f.

Thanks to Adam Larat for finding this bug.