Trac #31051: support /bin -> /usr/bin symlink in gp2c spkg-configure.
On systems where /bin is a symlink to /usr/bin, we can detect the "gp"
executable at /bin/gp, which isn't its real location. This then leads
to us looking for pari.cfg in the wrong place, and subsequently
failing to find it. This commit adds a special case to gp2c's autoconf
script to work around the snafu.