Added missing module and renamed test.
Run the performance test as: ./sage performance_test.sage

The output is CSV-formatted and  measures how much time various parts
of the code spends during the computation of one (homological) step of a
certain free resolution.  This computation is done one (internal) degree
at a time, and each line in the CSV file corresponds to a single degree.
I recording in the degree range 20-60.

Each line contains the following measurements:

"total time": This is the number of seconds it takes to complete one
step of the free resolution in one homogeneous degree.

"accounted time": The number of seconds which are accounted for (I only
measured the time spent in certain parts which I suspected was consuming
most).

"SteenrodAlgebra": The number of seconds spent calling the Steenrod
module.

"linear algebra": The number of seconds spent doing linear algebra.

For each line, the two last values should add up to "accounted time" (or
close to it.  I left out some more measurements which were tiny in
comparison.)