Further rework of 45b175e43621795f890204a9cdfb30b524f961fa which introduced some regressions.
PermutationGroup_generic._libgap is gone: We add a _libgap_
implementation (as a faster alternative to the existing _gap_init_)
and make it a cached_method.  If the user provides a gap_group argument
we simply use it to set the value of the cached_method.

The .gap() method now simply wraps ._libgap_().  To speed up creation
of SymmetricGroup and AlternatingGroup we also give them _libgap_
methods which are also cached as in the base class.