Trac #30668: don't run obsolete AC_PROG_CC_C99 macro in GCC check.
The AC_PROG_CC_C99 autoconf macro is obsolete and generates warnings
with >=autoconf-2.70. It has been "replaced" by AC_PROG_CC, which
now tests support for the various standards itself, and sets the
same output variables that e.g. AC_PROG_CC_C99 used to.

This commit drops one use of AC_PROG_CC_C99 from the GCC ./configure
check. That check still calls AC_PROG_CC, so support for C99 will
still be tested... but only if the user has a newer version of
autoconf. This choice is a trade-off that risks sacrificing a few
people with decades-old compilers, but eliminates a bit of future
cleanup work for the sage developers.