Trac #29617: compile a check for ECL_VERSION_NUMBER.
With the new version supported in ticket 31336, we are finally able to
support unpatched system installations of ecl. However, we can only
accept the *latest* version of ecl, because older ones are missing
critical patches.

This commit adds an AC_RUN_IFELSE() test to the ./configure script
that includes ecl/config.h and then returns whether or not

  ECL_VERSION_NUMBER < 210201

If the system version is too old, that will return "1" and the check
will fail. Otherwise, the entire program succeeds by default.