Trac #31395: eliminate the --memlimit option to the doctest runner.
Testing is currently impossible on Gentoo because the default memory
limit imposed by the doctest runner is too low for sage to function.
Rather than increase the limit once again, this commit removes the
memory limit entirely. This is not as huge of a change as it seems,
and the choice was made for three reasons:

  * Having a default limit occasionally breaks all testing on some
    platforms and requires the limit to be increased manually.

  * Setting a limit only works reliably on Linux where there are
    easier, more standard ways to set one (like ulimit -v).

  * There's only one test in the sage library that requires a memory
    limit to be set, and having one global memory limit would probably
    not be the correct approach if more tests requiring one were
    added.

This commit removes the --memlimit option, the corresponding code to
set a soft limit, and the support for the "optional - memlimit"
tag. The default memory limit was also thereby eliminated.  The one
test that requires a memory limit was refactored to set its own memory
limit on linux.

This reintroduces some boilerplate to the lone test that requires a
memory limit, but eliminates a recurring problem (the default limit
being too low) in the doctest runner at large. In the future, the
memory limit in that one test may need to be updated, but the absence
of a global limit should allow the doctest runner to function normally
up until that point.