Trac #30888: resolve $0 links in src/bin/sage.
When a symlink points to the final local/bin/sage, that script cannot
determine some important paths, and sage fails to start. This commit
copies & pastes the "resolvelinks" function from the top-level "sage"
script (that exists for much the same reason) into src/bin/sage where
it is now used on $0. When $0 us a symlink, we now obtain its target.

The code duplication is not ideal, but at the moment there is no other
straightforward way for local/bin/sage to determine where it lives;
this fixes the symlink issue until a better solution can be found.