Trac #31619: weaken the real/complex testing in is_positive_definite().
The is_positive_definite() method checks whether or not the base ring
of the given matrix is real or complex, but presently it rejects a few
inexact rings, such as RR and CC. This commit tweaks that test to
accept more inexact base rings.

At the moment this only changes errors like,

  ValueError: Could not see Real Field with 53 bits of precision
  as a subring of the real or complex numbers

into,

  TypeError: entries of the matrix must be in an exact ring, not
  Real Field with 53 bits of precision

because the underlying _indefinite_factorization() method rejects
inexact rings. That will be addressed separately.