Trac #31619: make the output of cholesky() more predictable.
The cholesky() method is currently basing its choice of a base ring
on the existence of a cached result; specifically on whether or not
the indefinite_factorization() was called with algorithm="symmetric"
or algorithm="hermitian". This is suboptimal in two ways:

  1. Things shouldn't break if somebody clears the cache.

  2. Computing the indefinite_factorization() with "hermitian" as
     the algorithm works just fine if you do it on a real matrix,
     but it will trick all subsequent calls to cholesky() into
     changing their results.

This commit updates the code (and related documentation) to return a
matrix over, in order of preference,

  (1) the fraction field of the base ring of the input matrix
  (2) the algebraic reals
  (3) the algebraic closure of field in (1).

It also recomputes the indefinite factorization if one is not cached.