Modify stdlibs/incs in misc/cython.py
On Homebrew, for example, python3-config doesn't remember the
libraries and includes for several packages.  This is so that
/usr/local doesn't bleed into other python projects.

This affects Sage at runtime since it can't locate packages
like gmp, etc.  To remedy this we loop through the standard
packages in misc/cython.py and if pkgconfig can find it
then we add the libraries and includes at the end of the
arguments.

This is a first attempt at fixing this problem.