Trac #31619: clean up documentation for is_positive_definite().
After some recent changes to the is_positive_definite() method, its
documentation refers to (and tests) a few things that aren't really
relevant any longer. This commit tries to clarify exactly what is
happening within the method, and streamlines the test cases to check
only those properties that are actually relevant to the method.