Trac #31619: replace is_positive_definite(certificate=True).
This optional argument has been replaced but is now deprecated in
favor of the user asking for the factorization separately. Moreover,
only the "L" and "d" portions of the block_ldlt() factorization are
returned as a certificate. This ignores the fact that a permutation
may be needed, but in most cases will agree with the old behavior, and
avoids breaking the interface an additional time (by returning a
triplet instead of the expected pair).