Added methodology for the deterministic calculation of the period matrix.
In order to allow for the efficient calculation of the period matrix, the method deterministic_line_integral is added which takes in bounding data as well as differentials and returns a value
which is rigorously bounded via known results. The necessary modifications are then made to period_matrix and matrix_of_integral_values in order to facilitate switching between these in an
efficient way (namely a way that allows for the period matrix to be cached).

In addition documentation changes are made. Because of dependencies on ticket #30698 and #32014 the alternate method cannot be checked yet, but by the configuration of the default parameters one
can still run period_matrix() and it is backwards compatible.