Trac #32037: update spkg-configure.m4 for lcalc-2.0.0.
Now that we need to detect lcalc >= 2.0.0, we can simplify the code
used to do so. The libLfunction API has changed in v2.0.0, so instead
of querying "lcalc --version" and parsing the result, this commit
modifies the test program to call the new API. If an old version of
lcalc (and thus of libLfunction) is present, the test program will
fail to compile.