Trac #29024: new spkg-configure.m4 for singular.
This new spkg-configure.m4 tries to use libSingular the way that we do
in sage/libs/singular/singular.pyx: by dlopen()ing the library with
some flags set. If we can guess the library's name, and if dlopen()
works, and if Singular is new enough... we'll accept the system copy,
and write the basename of the library into sage_conf. Otherwise, we
install the SPKG and write the old SAGE_LOCAL name into sage_conf.

The sage_conf variable will be used in a later commit to replace
the _get_shared_lib_path() hack currently in use.