Trac #32242: use GNU Info to obtain Singular docstrings.
Our Singular interface currently uses a jury-rigged parser to extract
docstrings from Singular's info file. This creates several maintenance
and performance issues that can now be avoided by using the optional
"info" SPKG.

This commit replaces the hand-written parser with a call to "info",
that (when all goes as planned) outputs the same string as before.
We also refactor the surrounding code to produce more consistent
docstrings, contained in a reST code block and with the same lead-in
sentence ("The Singular documentation for <foo> is given below.").

When GNU info is not installed, we now simply suggest that the user
install it. This is consistent with being unable to read Singular's
documentation from a singular_console() when Info is not installed.