Trac #32243: patch gfan to look for cddlib headers in new location.
The headers for cddlib are located in a "cddlib/" subdirectory these
days, but gfan is still only able to search for them in the top-level
include directory or in a subdirectory named "cdd/". Since SageMath
requires a newer version of cddlib that uses the new location, and
since there's no place to report these issues to gfan upstream, this
commit just patches the gfan sources in a few places to use the new
location.