Trac #32242: use fallback defaults for $INFOPATH.
The "info" from the GNU Info SPKG will not search the user's system
paths if his $INFOPATH variable is not set. In that case, we will miss
info files installed by system packages. To work around this, we
append a colon (:) to the $INFOPATH variable, which tells "info" to
search a default list of paths as last resorts. The full list is
defined as DEFAULT_INFOPATH in,

  https://git.savannah.gnu.org/cgit/texinfo.git/tree/info/filesys.h