Trac #8450: new fast-callable output wrapper for plotting.
Currently we are unable to plot expressions that involve complex
terms, even when the end result is real; take for example the
expression "abs(x*I)" that produces a TypeError when you try to plot
it. The error stems from the "domain" that we pass to fast_callable()
when we turn the expression into something that can be evaluated many
times quickly.

Since we desire a floating point result, we specify domain=float after
a fashion. However, this casts all of the *intermediate* expressions
to "float" as well -- problematic, after you substitute some value for
"x" into "x*I". To fix this issue, we would like to perform the
computations over CDF, but then cast the end result to float (or NaN,
if the result isn't real). To prepare for this, this commit adds a
FastCallablePlotWrapper class to wrap the function returned by
fast_callable(). When a FastCallablePlotWrapper is __call__ed, it
performs a complex-to-real conversion that is suitable for plotting.