close the internal pipe of multiprocessing.Process
Every finished, but not yet logged worker holds an open fd.
Thus when following a long running worker, so many finished tests can accumulate,
that the open files limit (ulimit -n) is reached.
This then causes the test suite to fail with 'OSError: [Errno 24] Too many open files'.

The open fd is due to a behavior change of the multiprocessing.Process class in python3.
It now opens a pipe internally, which it did not do in python2.

The solution is to call the close() method of multiprocessing.Process to close the internal pipe.
However, this method is only available since python 3.7.