Trac #21042: heuristic improvements for single-level contour plots.
Out plotting backend, matplotlib, doesn't grok contour plots whose
data all lie on one side of the sole contour. In particular, this
affects our implicit_function(), which works by plotting one "zero"
contour. For example,

  implicit_plot(lambda x,y: x-y, (-1,1), (-1,1))

works quite well; while

  implicit_plot(lambda x,y: abs(x-y), (-1,1), (-1,1))

which should be **the same plot**, complains that "No contour levels
were found within the data range," and produces a blank
page. Matplotlib does not seem too worried about this upstream, so
here we add some heuristics to detect the situation and produce
something a little better than an empty page.

In summary: if there's a single contour, and if all of the data lie on
one side of it, and if perturbing the data will change that, then we,

  1. Perturb the data by a tiny amount.
  2. Increase the number of plot points.
  3. Warn the user about what we did, and to be suspicious of
     the result.

In simple cases, you'll get something that looks reasonable, but the
quality of the result depends on how many of the perturbed points
happen to jump across the contour, and that's not easy to predict.
Regardless, the true value here is that we are able to warn the user,
who may then increase the number of plot points himself, or find some
other plotting strategy.