Trac #31275: drop qepcad-B-env.patch.
This patch sets two environment variables, $qe and $SINGULARPATH, if
they are not already set. However, no explanation for this is given.

The $qe variable is supposed to point to the "QEPCAD installation
directory," whatever that means. The default value for $qe from the
patch is "/usr/share/qepcad", and is apparently Fedora-specific, but
also somewhat irrelevant for us because we set a different value in
spkg-install.in.

The value of SINGULARPATH is also Fedora-specific, and is just plain
wrong for us, since Singular may not be installed as an SPKG.
Morever, SINGULARPATH isn't an environment variable that belongs to
QEPCAD -- it belongs to Singular (and does not need a default).

Since this patch is getting in the way of system Singular and no
rationale for it can be deduced, we remove it for now.