Trac #29024: put the shared library extension in sage_conf.
Preparing for a system copy of Singular, we need to be able to find
the name of the main shared library that Singular installs. In doing
so, we avoid the need to guess in _get_shared_lib_path().

In general, the name of this library will be libSingular.$SHLIBEXT,
where $SHLIBEXT represents the extension of shared libraries on the
system. As such, it makes sense to factor out the "name of the
library" into two parts: its base name, and $SHLIBEXT, which can be
determined independently and can be used with other libraries.

This commit uses an autoconf macro from gettext to determine the
shared library extension, and substitutes it into sage_conf.py as
SHLIBEXT.