Trac #31553: work around space symbols in maxima_lib's "retrieve" function.
Our custom "undoing_true_false_printing_patch" maxima patch is a
roadblock for using the system copy of maxima. The goal of the patch
is to eliminate some line breaks that appear unexpectedly when maxima
asks the user a question; so without the patch, we get, for example,

  Is x an even\ number?

This turns out to be caused by certain space characters being
represented by symbols rather than strings. This commit updates our
copy of maxima's "retrieve" function (which preprocesses said
questions) to (a) mirror the latest upstream implementation, and (b)
convert those symbols to strings. Afterwards, it removes the custom
patch that is no longer necessary.

Thanks to Nils Bruin for figuring out what the real problem was here.