Trac #32656: remove commented GAP code in quadratic_form__neighbors.py.
The sage/quadratic_forms/quadratic_form__neighbors.py file contains some
commented code for messing with the GAP memory pool size. The code is of
course never run, but since we intend to remove the GAP memory management
code altogether in a future commit, now is as good a time as any to prune
these dead comments.