Trac #32656: eliminate manual GAP memory management.
Our GAP interface provides a few functions to get/set the GAP memory
pool size. That size used when initializing the GAP interface, being
passed as the argument to GAP's "-o" and "-s" command-line flags:

  https://www.gap-system.org/Manuals/doc/ref/chap3.html

In preparation for the removal of psutil and sage.misc.getusage, we
remove all of this custom memory management code. The "-o" flag
specifies an upper bound, and can simply be omitted.

We also omit the "-s" flag in this commit, but that flag is a bit more
delicate, since it may affect users with custom GAP modules. However,
if the omission of "-s" causes a problem down the line, and if the GAP
documentation's assumptions about mmap et al. are correct, then we
can simply add back the "-s" flag with a large, static value.