Trac #32656: remove unused sage/modular/modform/test.py.
This file contains a bunch of vestigial modular forms tests that were
added to the tree in 2008, and which would (if ever run) make use of
sage.misc.getusage. Since this file essentially does nothing, we
remove it now so that e.g. `git grep getusage` does not show false
positives prior to the removal of that module.