Trac #32656: use static stack size bound for PARI.
Cypari requires that we pass both an initial stack size and a maximum
stack size to its Pari() constructor. Previously, the initial stack
size was chosen to be 1MB, and the maximum was set at 25% of whatever
the virtual_memory_limit() function returns.

In preparation for the removal of virtual_memory_limit(), we change
that upper limit to a static value of 1GiB. Since this is only a
bound, we should be able to increase it in the future if 1GiB proves
too low; I quote the docstring for Pari.__init__ from cypari2,

  When the PARI system is already initialized, the PARI stack is only
  grown if ``size`` is greater than the current stack...