Trac #32832: allow dict test to pass with unpatched cython.
We have a test in sage/cpython/dict_del_by_value.pyx that began to
fail when we switched to python3, and was fixed in trac 26855 by
patching cython. While the patch is a backport from cython-3.0, that
release has yet to materialize after several years, and now the need
for a patched cython is holding back efforts to modularize and package
sage.

This commit updates the test in question so that it works (and still
tests the same functionality) without a patched cython.