Trac #29665: set $SAGE_PIP_INSTALL_FLAGS for --enable-system-site-packages.
We need to pass some special flags to "pip install" to make it do the
right thing in the precence of system-site packages. Here we introduce
a new autoconf variable, SAGE_PIP_INSTALL_FLAGS, that will be used
to pass those flags to the build system.

With autoconf aware of the variable, we substitute its value into the
sage-build-env-config script for use in spkg-install.in.