Trac #32836: don't let libSingular clobber the user's PATH.
When we init_libsingular() in sage, singular prepends a few paths to
the user's PATH environment variable, one of them being the directory
where the "singular" executable lives. With a system singular package
installed, this has the unfortunate consequence that something like
"/usr/bin" will be given preference over the rest of your PATH, and
in particular, preference over $SAGE_LOCAL/bin.

This commit saves the PATH before calling init_libsingular(), and
restores it afterwards.