Trac #32511: fix argument order returned from minimize_constrained().
The minimize_constrained() function is supposed to return a minimizer
for its "func" argument, but it returns a tuple that corresponds to
the named arguments in alphabetical order, rather than in the order
that "func" expects them. For example, the minimizer of f(y,x) would
be returned as (x,y), because "x" comes before "y" alphabetically.

The root cause of this problem is the use of func.variables() to obtain
the list of arguments. This commit changes it to func.arguments(), which
returns them in the order that "func" is expecting.