Trac #13674: fix parent of sparse matrix sums.
Our generic sparse matrix class implements matrix addition, but casts
the result to its own generic type. In particular for its new sparse
RDF/CDF subclass, this will do the wrong thing, since the result will
bypass the (overridden) subclass methods and use the generic ones.

This commit addresses the problem by using the parent of the left
summand for the result. Somehow -- as a new test shows -- this even
does the right thing when the order of the summands is reversed.