Trac #32964: replace some uses of var() with SR.var().
This is unrelated to the main purpose of ticket 32964, but while I was
editing this file, I noticed that several examples use var() to inject
symbolic expressions into the current scope. This commit changes them
to use the preferred form `x = SR.var('x')` whose side-effects are
much easier to predict.