Trac #32973: increase the default test timeout to ten minutes.
This is intended to prevent timeouts on older machines where we now
have files like sage/manifolds/differentiable/tensorfield.py that take
over six minutes to test with a single thread. Since the timeout is a
failsafe to prevent infinite loops in the test suite, increasing it
should do no harm on newer machines. The worst thing that can happen
in the event of an infinite loop is that everyone will churn for a few
extra minutes before the process is killed.