Trac #33031: use superclass for naive RBF/CDF is_hermitian().
The Matrix_double_dense class has its own implementation of
is_hermitian() with an additional "algorithm" parameter. The "naive"
algorithm is now identical to the superclass implementation since
we've added a tolerance, so this commit replaces the duplicated code
with a super() call. The code for algorithm="orthonormal" has been
left alone and is still default, to avoid changing the behavior
of this class.