Trac #33058: don't repeat random tests in sage.graphs.generic_graph.
Several doctests in this module repeat random tests in a loop like,

  sage: for i in range(10):
  ....:     g = graphs.RandomGNP(n, 0.2)
  ....:     <test>

which generally just waste time after the code is working. Ultimately
the coverage from those tests should come from the number of machines
running the test suite, or from the --file-iterations=N parameter to
the doctest runner (if you want to force random tests to be run more
than once). This commit eliminates the loops so that the tests are
only run once per file iteration.