Trac #33100: switch default RDF/CDF is_hermitian algorithm to "naive".
The "naive" algorithm is the same one used in the superclass, and in
the is_symmetric() method -- it simply checks the entries against the
ones across the diagonal using a tolerance. But note that this does
not suffer from some terrible instability as the name would imply.

We make the switch in preparation for checking is_hermitian() within
the cholesky() method. There's at least one doctested example where
is_hermitian(algorithm="orthonormal") fails, and switching the default
to "naive" seems the simplest and most consistent way forward. This
may also allow us to eliminate the subclass is_hermitian() in the
future, since now the default behavior in the subclass is identical to
that of the superclass.