Trac #31424: add more precision for symbolic comparisons.
The test_relation() method of symbolic expressions is invoked when you
ask for a boolean answer from a symbolic relation like bool(x < y). If
the expression is not algebraic, the method tries to deduce an answer
over RIF or CIF which have 53 bits of precision by default. Two
examples on Trac 31424 and Trac 31665 show that this is not enough:
their answers are "I don't know," which is subsequently converted to
False by the boolean cast.

Ideally, users would use the test_relation() method directly, rather
than casting to bool, to determine the truthitude of these
expressions. However, we should not use documentation as an excuse for
returning a wrong answer when the right answer is available to us. I
have therefore increased the amount of precision in the method to 128
bits, which is enough to handle the two example problems.