Trac #33154: fix random failure in sage.plot.matrix_plot.
There's a random 3x3 polynomial matrix in this file that's supposed to
be "nonconstant," which in this context means that it contains at
least one polynomial entry of degree one or higher. But if you are
extremely unlucky, random_matrix() may generate you nine entries of
degree < 1. Here we avoid the issue by generating an additional
polynomial of degree >= 1, and inserting it into the matrix at a
random location.

with '#' will be ignored, and an empty message aborts the commit.  # #
On branch u/mjo/ticket/33154 # Changes to be committed: # modified:
src/sage/plot/matrix_plot.py # # Untracked files: # test.py #