Trac #33100: allow nonpositive tolerance for RDF/CDF is_hermitian().
Our RDF/CDF matrix subclass has a check to ensure that the tolerance
passed to is_hermitian() is strictly positive. The superclass does not
do that, and there's _technically_ no problem with using a negative
tolerance -- all comparisons will simply fail -- so we remove the
check here to simplify the class hierarchy a bit.

The eventual goal is to consolidate and unify the API and behavior of
all these methods.