Trac #33213: replace SAGE_TMP in pexpect interface.
This consists mainly of replacing a SAGE_TMP_INTERFACE location with a
tempfile.NamedTemporaryFile() in _local_tmpfile(). The old path,
however, was deterministic for a given PID; the new path, being
random, is not. This necessitates another change: we can't call
self.quit() unconditionally in self._start(), because that indirectly
causes a new temporary file (with a new name!) to be generated. This
happened to work before only because the new name and the old name
would be the same. In any case, it is fixed by calling quit() only
if the interface is_running().