Trac #33213: replace ECL_TMP (based on SAGE_TMP) in maxima/ecl interfaces.
We have historically massaged TMPDIR and TMP when interfacing with ECL
to ensure that SAGE_TMP was used consistently for temporary
files. Simply removing this code seems to do the trick, since ECL
wants to use the system location by default.

The existing ECL_TMP path is marked with a deprecation warning.