Trac #33213: move sage.interfaces.cleaner.cleaner() to quit interface.
With sage-cleaner on the chopping block, its interface is going to
need to be removed eventually. Part of it is still used: our expect
interface allows instances to register themselves to be killed with
the expect_quitall() function in sage.interfaces.quit. To keep that
working, we move/rename the function that registers a a process to
register_spawned_process() in the module sage.interfaces.quit.