Trac #33213: use @cached_function for the spawned processes file.
This fixes an accidental change in behavior that was introduced in an
earlier commit. The path to this file shouldn't be computed when the
module is loaded -- only when it is requested. A @lazy_string would
work here, but in my opinion the added abstraction on top of the
more-familiar @cached_function is not worth it.