Trac #33213: ensure temporary gap workspace is created in a doctest.
We have a GAP interface test that temporarily switches the workspace
location and creates an instance of Gap(), which is supposed to
(indirectly) reset the workspace. However, this doesn't necessarily
happen: there's a global variable called first_try that needs to be
True. Here we tweak the doctest to ensure that it is.