Trac #32867: fuzz some maxima test output to support system installs.
We have one maxima nintegral() test that fails when the system
"maxima" uses a 32-bit SBCL (as opposed to ECL). This is due to a bug
in SBCL, but the error is minescule -- not something we should be
testing in the first place.

This patch adds some "..." to ignore the unstable digits. Thanks to
Gonzalo Tornaría for noticing and fixing the problem.