Make WordMorphism str use word separator if long
This commit makes WordMorphism's string representation use whatever the
current word separator is if the values in the codomain have length
larger than 1. This is to prevent ambiguous representations such as

    sage: a = WordMorphism({1: [1, 2], 2: [12], 12:[1, 2]})
    sage: print(a)

Currently, this prints

    print(a)  # old
    1->12, 12->12, 2->12

And two of the "12"s above are actually `1, 2`. With this commit, it
will print

    print(a)  # new
    1->1,2, 12->1,2, 2->12

This maintains the ability to use different separators besides commas.
For example

    sage: WordOptions(letter_separator=":")
    print(a)
    1->1:2, 12->1:2, 2->12