Trac #33134: add random (disabled) test for hilbert_numerator() algorithms.
In a previous commit on Trac 33134, a test checking for incorrect
hilbert_numerator() output was deleted. To replace it, this commit
adds a random test that ensures both algorithms for
hilbert_numerator() agree. However, the test is disabled for now,
until the bug that caused the incorrect output can be fixed.