Change how the 'algorithm' keyword behaves.
Previously, the 'algorithm' keyword only mattered for WittRing_p_typical. All the others would automatically
select an algorithm. Also, WittRing_base didn't even acecpt it in the constructor. Now, WittRing_finite_field
and WittRing_p_invertible are the only ones that ignore it. We also introduce the 'auto' choice as the
default,
wherein the code tries to select the fastest algorithm.

In the future, we may change that WittRing_finite_field and WittRing_p_invertible ignore this argument.